face_recognition/requirements.txt
flask==3.0.3
flask-cors==4.0.1
face-recognition==1.3.0
opencv-python==4.10.0.84
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
filelock==3.15.4
